requirements.txt
rich>=13.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
prompt_toolkit>=3.0.0
questionary>=2.0.0
duckduckgo-search>=6.0.0
tavily-python>=0.5.0
